tests/wpt/tests/css/css-inline/animation/initial-letter-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<link rel=help href="https://drafts.csswg.org/css-transitions-2/#transition-property-property">
<link rel=help href="https://drafts.csswg.org/css-inline/#sizing-drop-initials">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'initial-letter',
  from: 'initial',
  to: '123'
});
</script>
